+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to the
 [Haskell Package Versioning Policy](https://pvp.haskell.org/).
 
 ## Unreleased
+- change LCTRS (format ..) tag syntax to include ':standard'. See:
+  http://project-coco.uibk.ac.at/ARI/lctrs.php
 
 ## 0.5.0.0 - 2023-12-15
 - many small fixes to the ARI parser